+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next
+- **Added:** Support for `itemProps` in `NestField`. [\#181](https://github.com/vazco/uniforms/issues/181)
 - **Fixed:** Handling of arrays in SimpleSchema@2. [\#200](https://github.com/vazco/uniforms/issues/200)
 
 ## [v1.11.3](https://github.com/vazco/uniforms/tree/v1.11.3) (2017-02-16)